refactor: Migrate compress generator to typescript (LPS-119618)
