+++ src/compress/index.ts
@@ -1,28 +1,19 @@
-const {
-  ADD_DEPLOYMENT_DESCRIPTOR_DEFAULT,
-  ADD_DEPLOYMENT_DESCRIPTOR_MESSAGE,
+import {
   ADD_DEPLOYMENT_DESCRIPTOR_VAR,
-  DEPLOYMENT_DESCRIPTOR_COMPANY_DEFAULT,
-  DEPLOYMENT_DESCRIPTOR_COMPANY_MESSAGE,
   DEPLOYMENT_DESCRIPTOR_COMPANY_VAR,
-  DEPLOYMENT_DESCRIPTOR_GROUP_DEFAULT,
-  DEPLOYMENT_DESCRIPTOR_GROUP_MESSAGE,
   DEPLOYMENT_DESCRIPTOR_GROUP_VAR,
-} = require('../utils/constants');
-const { default: CustomGenerator } = require('../utils/custom-generator');
-const compress = require('./compress');
+} from '../utils/constants';
+import CustomGenerator from '../utils/custom-generator';
+import compress from './compress';
 
-module.exports = class extends CustomGenerator {
-  /**
-   * @inheritdoc
-   */
-  async prompting() {
+export default class extends CustomGenerator {
+  async asking(): Promise<void> {
     await this.ask([
       {
         type: 'confirm',
         name: ADD_DEPLOYMENT_DESCRIPTOR_VAR,
-        message: ADD_DEPLOYMENT_DESCRIPTOR_MESSAGE,
-        default: ADD_DEPLOYMENT_DESCRIPTOR_DEFAULT,
+        message: 'Add deployment descriptor?',
+        default: true,
         when: !this.hasValue(ADD_DEPLOYMENT_DESCRIPTOR_VAR),
       },
     ]);
@@ -32,34 +23,28 @@ module.exports = class extends CustomGenerator {
         {
           type: 'input',
           name: DEPLOYMENT_DESCRIPTOR_COMPANY_VAR,
-          message: DEPLOYMENT_DESCRIPTOR_COMPANY_MESSAGE,
-          default: DEPLOYMENT_DESCRIPTOR_COMPANY_DEFAULT,
+          message: 'Deployment descriptor company Web ID?',
+          default: 'liferay.com',
           when: !this.hasValue(DEPLOYMENT_DESCRIPTOR_COMPANY_VAR),
         },
       ]);
 
       if (
         this.answers[DEPLOYMENT_DESCRIPTOR_COMPANY_VAR] &&
-        this.answers[DEPLOYMENT_DESCRIPTOR_COMPANY_VAR] !== '' &&
         this.answers[DEPLOYMENT_DESCRIPTOR_COMPANY_VAR] !== '*'
       ) {
         await this.ask([
           {
             type: 'input',
             name: DEPLOYMENT_DESCRIPTOR_GROUP_VAR,
-            message: DEPLOYMENT_DESCRIPTOR_GROUP_MESSAGE,
-            default: DEPLOYMENT_DESCRIPTOR_GROUP_DEFAULT,
+            message: 'Deployment descriptor group key?',
+            default: 'Guest',
             when: !this.hasValue(DEPLOYMENT_DESCRIPTOR_GROUP_VAR),
           },
         ]);
       }
     }
-  }
 
-  /**
-   * @inheritdoc
-   */
-  async writting() {
     await compress(this.destinationPath(), {
       [ADD_DEPLOYMENT_DESCRIPTOR_VAR]: this.getValue(
         ADD_DEPLOYMENT_DESCRIPTOR_VAR
@@ -72,4 +57,4 @@ module.exports = class extends CustomGenerator {
       ),
     });
   }
-};
+}
