+++ src/import/__tests__/import.spec.js
@@ -1,6 +1,6 @@
 const JSZip = require('jszip');
 
-const compress = require('../../compress/compress');
+const { default: compress } = require('../../compress/compress');
 const api = require('../../utils/api');
 const { ADD_DEPLOYMENT_DESCRIPTOR_VAR } = require('../../utils/constants');
 const getTestFixtures = require('../../utils/get-test-fixtures');
