+++ src/utils/constants.js
@@ -1,5 +1,3 @@
-module.exports.ADD_DEPLOYMENT_DESCRIPTOR_DEFAULT = true;
-module.exports.ADD_DEPLOYMENT_DESCRIPTOR_MESSAGE = 'Add deployment descriptor?';
 module.exports.ADD_DEPLOYMENT_DESCRIPTOR_VAR = 'addDeploymentDescriptor';
 
 module.exports.ADD_SAMPLE_CONTENT_DEFAULT = true;
@@ -25,14 +23,8 @@ module.exports.COLLECTION_SLUG_VAR = 'collectionSlug';
 module.exports.DATA_LFR_SUPPORTED = 'dataLfrSupported';
 module.exports.DATA_LFR_SUPPORTED_MIN_VERSION = '7.3.3';
 
-module.exports.DEPLOYMENT_DESCRIPTOR_COMPANY_DEFAULT = 'liferay.com';
-module.exports.DEPLOYMENT_DESCRIPTOR_COMPANY_MESSAGE =
-  'Deployment descriptor company Web ID?';
 module.exports.DEPLOYMENT_DESCRIPTOR_COMPANY_VAR = 'companyWebId';
 
-module.exports.DEPLOYMENT_DESCRIPTOR_GROUP_DEFAULT = 'Guest';
-module.exports.DEPLOYMENT_DESCRIPTOR_GROUP_MESSAGE =
-  'Deployment descriptor group key?';
 module.exports.DEPLOYMENT_DESCRIPTOR_GROUP_VAR = 'groupKey';
 
 module.exports.FRAGMENT_COLLECTION_SLUG_MESSAGE = 'Choose a collection';
